Add source code practice 10.
