+++ BaiTap10/TuanDQ/AutoLayouts/AutoLayouts/Main/ApplicationViewController.swift
@@ -1,0 +1,30 @@
+//
+//  ApplicationViewController.swift
+//  AutoLayouts
+//
+//  Created by MBA0051 on 10/7/19.
+//  Copyright © 2019 MBA0051. All rights reserved.
+//
+
+import UIKit
+
+class ApplicationViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
