+++ BaiTap10/TuanDQ/AutoLayouts/AutoLayouts/PracticeOne/CenterSquareAutosize/CenterSquareAutoSizeViewController.xib
@@ -7,7 +7,7 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="RectangleAutoSizeViewController" customModule="AutoLayouts" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="CenterSquareAutoSizeViewController" customModule="AutoLayouts" customModuleProvider="target">
             <connections>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
@@ -17,15 +17,22 @@
             <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
-                <view contentMode="scaleToFill" fixedFrame="YES" translatesAutoresizingMaskIntoConstraints="NO" id="XYx-qB-5WO">
-                    <rect key="frame" x="49" y="129" width="317" height="638"/>
-                    <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                    <color key="backgroundColor" systemColor="linkColor" red="0.0" green="0.47843137250000001" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="qvQ-vG-Wrj">
+                    <rect key="frame" x="87" y="333" width="240" height="240"/>
+                    <color key="backgroundColor" systemColor="systemGreenColor" red="0.20392156862745098" green="0.7803921568627451" blue="0.34901960784313724" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                    <constraints>
+                        <constraint firstAttribute="width" constant="240" id="faF-CZ-cam"/>
+                        <constraint firstAttribute="width" secondItem="qvQ-vG-Wrj" secondAttribute="height" multiplier="1:1" id="gi1-RB-QWX"/>
+                    </constraints>
                 </view>
             </subviews>
             <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <constraints>
+                <constraint firstItem="qvQ-vG-Wrj" firstAttribute="centerY" secondItem="fnl-2z-Ty3" secondAttribute="centerY" id="EAz-Ei-fiu"/>
+                <constraint firstItem="qvQ-vG-Wrj" firstAttribute="centerX" secondItem="fnl-2z-Ty3" secondAttribute="centerX" id="FwP-XZ-VgX"/>
+            </constraints>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
-            <point key="canvasLocation" x="131.8840579710145" y="126.5625"/>
+            <point key="canvasLocation" x="133" y="108"/>
         </view>
     </objects>
 </document>
