+++ BaiTap10/TuanDQ/AutoLayouts/AutoLayouts/PracticeOne/PracticeOneViewController.swift
@@ -1,0 +1,27 @@
+//
+//  PracticeOneViewController.swift
+//  AutoLayouts
+//
+//  Created by MBA0051 on 10/7/19.
+//  Copyright © 2019 MBA0051. All rights reserved.
+//
+
+import UIKit
+
+class PracticeOneViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+        title = "Center square"
+    }
+    @IBAction func squareCenterClick(_ sender: UIButton) {
+        if (sender.tag == 1) {
+            navigationController?.pushViewController(CenterSquareViewController(), animated: true)
+        } else {
+            navigationController?.pushViewController(CenterSquareAutoSizeViewController(), animated: true)
+        }
+    }
+    
+}
